scratch_track/requirements.txt
asgiref==3.2.10
Django==3.0.7
django-crispy-forms==1.9.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2020.1
sqlparse==0.3.1
